split generate script into separate files
